requirements.txt
pandas
geopandas
shapely
factor-analyzer
scikit-learn
matplotlib
seaborn
plotly
dash
streamlit
streamlit_folium
folium
firebase-admin
python-dotenv
geopy
